GlaucomaM {TH.data}	R Documentation

Glaucoma Database

Description

The GlaucomaM data has 196 observations in two classes. 62 variables are derived from a confocal laser scanning image of the optic nerve head, describing its morphology. Observations are from normal and glaucomatous eyes, respectively.

Usage

data("GlaucomaM")

Format

This data frame contains the following predictors describing the morphology of the optic nerve head and a membership variable:

ag

area global.

at

area temporal.

as

area superior.

an

area nasal.

ai

area inferior.

eag

effective area global.

eat

effective area temporal.

eas

effective area superior.

ean

effective area nasal.

eai

effective area inferior.

abrg

area below reference global.

abrt

area below reference temporal.

abrs

area below reference superior.

abrn

area below reference nasal.

abri

area below reference inferior.

hic

height in contour.

mhcg

mean height contour global.

mhct

mean height contour temporal.

mhcs

mean height contour superior.

mhcn

mean height contour nasal.

mhci

mean height contour inferior.

phcg

peak height contour.

phct

peak height contour temporal.

phcs

peak height contour superior.

phcn

peak height contour nasal.

phci

peak height contour inferior.

hvc

height variation contour.

vbsg

volume below surface global.

vbst

volume below surface temporal.

vbss

volume below surface superior.

vbsn

volume below surface nasal.

vbsi

volume below surface inferior.

vasg

volume above surface global.

vast

volume above surface temporal.

vass

volume above surface superior.

vasn

volume above surface nasal.

vasi

volume above surface inferior.

vbrg

volume below reference global.

vbrt

volume below reference temporal.

vbrs

volume below reference superior.

vbrn

volume below reference nasal.

vbri

volume below reference inferior.

varg

volume above reference global.

vart

volume above reference temporal.

vars

volume above reference superior.

varn

volume above reference nasal.

vari

volume above reference inferior.

mdg

mean depth global.

mdt

mean depth temporal.

mds

mean depth superior.

mdn

mean depth nasal.

mdi

mean depth inferior.

tmg

third moment global.

tmt

third moment temporal.

tms

third moment superior.

tmn

third moment nasal.

tmi

third moment inferior.

mr

mean radius.

rnf

retinal nerve fiber thickness.

mdic

mean depth in contour.

emd

effective mean depth.

mv

mean variability.

Class

a factor with levels glaucoma and normal.

Details

All variables are derived from a laser scanning image of the eye background taken by the Heidelberg Retina Tomograph. Most of the variables describe either the area or volume in certain parts of the papilla and are measured in four sectors (temporal, superior, nasal and inferior) as well as for the whole papilla (global). The global measurement is, roughly, the sum of the measurements taken in the four sector.

The observations in both groups are matched by age and sex to prevent any bias.

Source

Torsten Hothorn and Berthold Lausen (2003), Double-Bagging: Combining classifiers by bootstrap aggregation. Pattern Recognition, 36(6), 1303–1309.

[Package TH.data version 1.1-1 Index]