Westbc {TH.data}	R Documentation

Breast Cancer Gene Expression

Description

Gene expressions for 7129 genes in 49 breast cancer samples and the status of lymph node involvement.

Usage

data("Westbc")

Format

An list with two elements to be converted to class ExpressionSet (see package Biobase).

Details

A full description of the data can be found in West et al. (2001) and an application of boosted linear models is given by Buehlmann (2006).

Source

Mike West, Carrie Blanchette, Holly Dressman, Erich Huang, Seiichi Ishida, Rainer Spang, Harry Zuzan, John A. Olson Jr., Jeffrey R. Marks and Joseph R. Nevins (2001), Predicting the clinical status of human breast cancer by using gene expression profiles, Proceedings of the National Academy of Sciences, 98, 11462-11467.

References

Peter Buehlmann (2006), Boosting for high-dimensional linear models. The Annals of Statistics, 34(2), 559–583.

Peter Buehlmann and Torsten Hothorn (2007), Boosting algorithms: regularization, prediction and model fitting. Statistical Science, 22(4), 477–505.

Examples

  ## Not run: 
    library("Biobase")
    data("Westbc", package = "TH.data")
    westbc <- new("ExpressionSet",
          phenoData = new("AnnotatedDataFrame", data = Westbc$pheno),
          assayData = assayDataNew(exprs = Westbc$assay))
  
## End(Not run)

[Package TH.data version 1.1-1 Index]