birds {TH.data}	R Documentation

Habitat Suitability for Breeding Bird Communities

Description

Environmental variables and bird counts for identifying suitable bird habitats

Usage

data("birds")

Format

A data frame with 258 observations on the following 10 variables.

GST

Growing stock per grid

DBH

Mean diameter of the largest three trees

AOT

Age of oldest tree

AFS

Age of forest stand

DWC

Amount of dead wood of conifers

LOG

Amount of logs per grid

x_gk

grid location, x coordinate

y_gk

grid location, y coordinate

SG4

observed number of birds from structural gild 4: Requirement of regeneration (Phylloscopus trochilus, Aegithalos caudatus)

SG5

observed number of birds from structural gild 5: Requirement of regeneration combined with planted conifers (Phylloscopus collybita, Turdus merula, Sylvia atricapilla).

Details

Counts of breeding bird communities collected at 258 observation plots in a northern Bavarian forest district are the response variable of interest. Along with the number of birds in two structural gilds, 6 covariates are given here and one is interested in quantifying their impact on habitat suitability.

Source

Joerg Mueller (2005). Forest structures as key factor for beetle and bird communities in beech forests. PhD thesis, Munich University of Technology. http://mediatum.ub.tum.de

References

Thomas Kneib and Joerg Mueller and Torsten Hothorn (2008), Spatial smoothing techniques for the assessment of habitat suitability, Environmental and Ecological Statistics, 15(3), 343–364.

[Package TH.data version 1.1-1 Index]