mammoexp {TH.data}	R Documentation

Mammography Experience Study

Description

Data from a questionaire on the benefits of mammography.

Usage

data(mammoexp)

Format

A data frame with 412 observations on the following 6 variables.

ME

Mammograph experience, an ordered factor with levels Never < Within a Year < Over a Year

SYMPT

Agreement with the statement: ‘You do not need a mamogram unless you develop symptoms.’ A factor with levels Strongly Agree, Agree, Disagree and Strongly Disagree

PB

Perceived benefit of mammography, the sum of five scaled responses, each on a four point scale. A low value is indicative of a woman with strong agreement with the benefits of mammography.

HIST

Mother or Sister with a history of breast cancer; a factor with levels No and Yes.

BSE

Answers to the question: ‘Has anyone taught you how to examine your own breasts?’ A factor with levels No and Yes.

DECT

Answers to the question: 'How likely is it that a mammogram could find a new case of breast cancer?' An ordered factor with levels Not likely < Somewhat likely < Very likely.

Source

Hosmer and Lemeshow (2000). Applied Logistic Regression, 2nd edition. John Wiley & Sons Inc., New York. Section 8.1.2, page 264.

[Package TH.data version 1.1-1 Index]