mn6.9 {TH.data}	R Documentation

I.Q. and attitude towards science

Description

Responses given by 2982 New Jersey high-school seniors on 4 questions concerning attitude towards science. Also recorded was whether students had a high or low I.Q.

Usage

data(mn6.9)

Format

A data frame with 2982 observations on the following 5 variables.

y1

Agree=1/disagree=0 to "The development of new ideas is the scientist's greatest source of satisfaction"

y2

Agree=1/disagree=0 to "Scientists and engineers should be eliminated form the military draft"

y3

Agree=1/disagree=0 to "The scientist will make his maximum contribution to society when he has freedom to work on problems that interest him"

y4

Agree=1/disagree=0 to "The monetary compensation of a Nobel Prize-winner in physics should be at least equal to that given to popular entertainers"

group

I.Q. levels: 1=low, 2=high

Source

McCullagh, P. and Nelder, J.A. (1989, p. 239). Generalized Linear Models. Second Edition. Chapman & Hall/CRC.

copied from multmod package 1.0 (CRAN archive)

[Package TH.data version 1.1-1 Index]