sphase {TH.data}	R Documentation

S-phase Fraction of Tumor Cells

Description

S-phase fraction of tumor cells in breast cancer patients.

Usage

data("sphase")

Format

This data frame contains the following columns:

SPF

S-phase fraction

RFS

recurrence free survival

event

censoring indicator: FALSE means censored, TRUE is an event.

Details

The data have been used to address the question whether a simple cutpoint in S-phase fraction can be used to discriminate between patients with good and bad prognosis (for example in Hothorn & Lausen, 2003).

Source

J. Pfisterer, F. Kommoss, W. Sauerbrei, D. Menzel, M. Kiechle, E. Giese, M. Hilgarth & A. Pfleiderer (1995). DNA flow cytometry in node positive breast cancer: Prognostic value and correlation to morphological and clinical factors. Analytical and Quantitative Cytology and Histology 7(6), 406–412.

References

Torsten Hothorn & Berthold Lausen (2003). On the Exact Distribution of Maximally Selected Rank Statistics. Computational Statistics & Data Analysis 43(2), 121–137.

[Package TH.data version 1.1-1 Index]